Add files via upload
Code works but the performance metrics are very low. I am still playing around with RF parameters.
I am also getting the following warnings when run LOOCV: 
"FutureWarning: Using a non-tuple sequence for multidimensional indexing is deprecated; use `arr[tuple(seq)]` instead of `arr[seq]`. In the future this will be interpreted as an array index, `arr[np.array(seq)]`, which will result either in an error or a different result."
"F-score is ill-defined and being set to 0.0 in labels with no true samples." - same for Specificity, Precision and Recall